moved loadYAML to loading.go
